fix 'make' when using glibc >= 2.28
glibc >= 2.28 supports `statx` out of the box. This means we get an
error when attempting to build `liburing` when glibc is >= 2.28, for
e.g. in my glibc version 2.33 I get the following errors

```
:26:8: error: redefinition of ‘struct statx_timestamp’
   26 | struct statx_timestamp
      |        ^~~~~~~~~~~~~~~
In file included from statx.c:14:
```
or
```
error: redefinition of ‘struct statx’
   30 | struct statx
      |        ^~~~~
In file included from statx.c:14:
```
This commit fixes 'make' error by choosing to use the glibc version
of `statx` if available.

glibc 2.28 release notes:
https://sourceware.org/git/?p=glibc.git;a=blob;f=NEWS;h=154ab22d7ca065af5233406927302bb7f6a66890;hb=3c03baca37fdcb52c3881e653ca392bba7a99c2b#l84

Signed-off-by: Bikal Lem <gbikal+git@gmail.com>